+++ ui/packages/catalog-ui-search/src/main/webapp/react-component/route/index.tsx
@@ -12,4 +12,4 @@
  * <http://www.gnu.org/licenses/lgpl.html>.
  *
  **/
-export { default } from './route-container'
+export { default } from './route'
